abstract ClosurePrompts
